refactored welcome fragment
