+++ app/src/main/java/edu/us/ischool/bchong/info448project/MainGameActivity.kt
@@ -5,6 +5,7 @@ import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.activity_maingame.*
 
 class MainGameActivity : AppCompatActivity() {
